Migrate build system to gradle.
Also begin introducing standard formats and protobufs
