+++ settings.gradle.kts
@@ -1,0 +1,6 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ */
+
+rootProject.name = "data-tools"
+include("proto")
