+++ src/main/kotlin/com/codality/data/tools/config/DbConf.kt
@@ -1,0 +1,6 @@
+package com.codality.data.tools.config
+
+data class DbConf (
+    val mongo: MongoConf? = null,
+    val redis: RedisConf? = null
+)
