+++ src/main/kotlin/com/codality/data/tools/config/FormatConf.kt
@@ -1,0 +1,6 @@
+package com.codality.data.tools.config
+
+data class FormatConf (
+    val csv: CsvConf? = null,
+    val json: JsonConf? = null
+)
