+++ src/main/kotlin/com/codality/data/tools/config/JsonConf.kt
@@ -1,0 +1,5 @@
+package com.codality.data.tools.config
+
+data class JsonConf (
+    private val parseNested: Boolean? = null
+)
