+++ src/main/kotlin/com/codality/data/tools/config/MongoConf.kt
@@ -1,0 +1,8 @@
+package com.codality.data.tools.config
+
+class MongoConf (
+    val host: String? = null,
+    val port: Int? = null,
+    val dbName: String? = null,
+    val workerCount: Int? = null
+)
