+++ src/main/kotlin/com/codality/data/tools/config/ProcessorConf.kt
@@ -1,0 +1,6 @@
+package com.codality.data.tools.config
+
+data class ProcessorConf (
+    val format: FormatConf? = null,
+    val db: DbConf? = null
+)
