+++ src/main/kotlin/com/codality/data/tools/config/RedisConf.kt
@@ -1,0 +1,7 @@
+package com.codality.data.tools.config
+
+class RedisConf (
+        val host: String? = null,
+        val port: Int? = null,
+        val workerCount: Int? = null
+)
