+++ src/test/resources/client.test.properties
@@ -11,4 +11,6 @@ mongo.database.name=testDb
 # Metadata
 csv.file=bin/data/movies_metadata.csv
 csv.regex=(?=[^\\s]),(?=[^\\s])
+csv.fields="id","title","budget","genres","popularity","companies","date","revenue"
 csv.has.header=true
+json.parseNested=true
